requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
asgiref==3.8.1
astroid==3.3.10
async-timeout==5.0.1
attrs==25.3.0
autopep8==2.3.2
backports.tarfile==1.2.0
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
cssselect==1.3.0
decorator==5.2.1
dicttoxml==1.7.16
dill==0.4.0
Django==5.2.1
django-ipware==7.0.1
docutils==0.21.2
exceptiongroup==1.3.0
frozenlist==1.6.0
future==1.0.0
geocoder==1.38.1
geoip2==5.1.0
h11==0.16.0
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
ip2geotools==0.1.6
IP2Location==8.10.5
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
lazy-object-proxy==1.11.0
lxml==5.4.0
markdown-it-py==3.0.0
maxminddb==2.7.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.7.0
multidict==6.4.4
nh3==0.2.21
outcome==1.3.0.post0
packaging==25.0
pip-review==1.3.0
pkginfo==1.12.1.2
platformdirs==4.3.8
propcache==0.3.1
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
Pygments==2.19.1
pylint==3.3.7
pyparsing==3.2.3
pyquery==2.0.1
PySocks==1.7.1
python-ipware==3.0.0
ratelim==0.1.6
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
SecretStorage==3.3.3
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
twine==6.1.0
typed-ast==1.5.5
typing==3.7.4.3
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.22.0
django-crontab
django-crispy-forms
django-qr-code